tests: Add generate_html_outputs test-suite
Add the `generate_html_outputs` test-suite with the following test:
  - definitely_lost_report

This test runs the `memcheck_runner.sh` script on the
`definitely_lost` test bin.
Then generates the HTML report using the `generate_html_report.sh`
script.

And finally compares the output report dir with a reference directory.

In order to compare results, the memcheck files are anonymized.
This process does the following changes to the file content:
  - Replaces each line's `==<process id>==` prefix with `==1==`
  - Replaces parent process id with `1`
  - Removes the absolute path part before the `memcheck-cover` dir.