generate_html_report.sh: Add an hide functionnality to the results
The generated report now has a javascript file.
It defines the ToogleAnalysisResultVisibility function, which toogle
the visibility of a valgrind result.

By default, each result is hidden. The viewer can click on the result
title in the report to show the result.
Clicking again will hide the result back.

So the used understand the title can be clicked, the following has
been added to the css stylesheet:
  - hovering the title will change the mouse cursor to a pointing hand
  - hovering the title will change the background color of the title

The arrow next to the title will point to the right if the result is
hidden. It will point down if the report is visible.