generate_html_report.sh: Add a result status to the analysis title
If the memcheck result is successful, the [SUCCESS] prefix is added.
If the memcheck result contains leaks, the [ERROR] prefix is added.

Added the `no_error_report` test to the `generate_html_outputs`
test-suite. This tests the `true` binary with dummy parameters.
It's output is a success memcheck report.
It also serves as a `cmd with params` check.