generate_html_report.sh: Add uninitialized value warning
Uninitialized value errors are now highlighted in orange.
This kind of error is considered a warning for now, criticality of
each error kind should be parameterizable in the future.

Added the following test to the `generate_html_outputs` testsuite:
  - error_report(uninitialized_value)

This test check the following errors:
  - Conditional jump or move depends on uninitialised value
  - Use of uninitialised value of size

With the following context types:
  - Uninitialised value was created by a heap allocation
  - Uninitialised value was created by a stack allocation