generate_html_report.sh: Add result asynchronous loading
If the server is a bit slow or the results contains many errors, the
page loading can be a bit slow and unresponsive.

To avoid this, the valgrind result in not embedded within the html
anymore. A javascript function now wrap each html part, which is
asynchronously loaded. Once each part gets loaded, it replaces the
report div content with the actual valgrind result.
Until then, a loader is displayed within the div.