generate_html_report.sh: Add valgrind's warnings support
From time to time, Valgrind raises warnings. For now, the following
ones are handled:
  - Warning: bad signal number 65 in sigaction()
  - Warning: invalid file descriptor -1 in syscall write()
  - Warning: ignored attempt to set SIGSTOP handler in sigaction();
             the SIGSTOP signal is uncatchable
  - Warning: ignored attempt to set SIGRT32 handler in sigaction();
             the SIGRT32 signal is used internally by Valgrind

Those messages are displayed in Orange, bold, underlined.
The report account those as warnings.

Their criticality is not parameterizable, it will always be 'warning'.

The following related test was added to the generate_html_outputs
test-suite:
  - error_report(valgrind_warnings,default_criticality)

---

Added the following function to the test utils:
  - expect_empty_file