generate_html_report.sh: Add some 'unadressable byte(s)' violations
Added support for the following violations:
  - Syscall param .* points to unaddressable byte(s)
  - contains unaddressable byte(s)

The later does not seem to be occuring anymore as of valgring 3.15.0,
since no test expects this error. However, older versions of such
tests did contain this errors.

---

Added the following related tests to the `generate_html_outputs`
test-suite:
  - error_report(unaddressable_bytes,default_criticality)
  - error_report(unaddressable_bytes,all_warnings_criticality)
  - error_report(unaddressable_bytes,all_errors_criticality)

The 'contains unaddressable byte(s)' error is emulated in the
ts_setup.sh.