generate_html_report.sh: Support generated suppressions in HTML report
The suppressions generated by Valgrind are now encapsulated into an
hidden div.
A button is displayed on top of each one to show or hide back the div.

A light yellow background color and a dotted border are used for both
the title and the content divs.

---

Added the related `violation_suppression_report` test to the
`generate_html_outputs` test-suite.