generate_html_report.sh: Support path without trailing slash
Passing a path without an ending slash could result in incorrect
index.html generation.

If the provided path does not end with a slash, it's now added.

---

Added the following related tests to the `generate_html_outputs`
test-suite:
  - relative_path_report(relative_path)
  - relative_path_report(relative_path_no_leading_slash)
  - relative_path_report(current_dir_in_path)
  - relative_path_report(current_dir_in_path_no_leading_slash)
  - relative_path_report(current_dir_out_path)
  - relative_path_report(current_dir_out_path_no_leading_slash)

This test-suite covers both relative paths and non-slash ending paths as
parameter values.

The special "." and ".." values are tested as well.

---

This commit fixes #7