tests: Force colored output in GitHub-CI
GitHub-CI does not run the tests in an interactive shell.
Thus colors are disabled.

Setting the memcheck_cover_always_use_colors variable to true will now
enable colors even within a non-interactive environnement.

This variable has been set to true if the GITHUB_RUN_ID is defined and
not empty for test scripts.