generate_html_report.sh: Add support to path prefix substitution
It's now possible to use the memcheck_path_prefix_replacement
associative array in the config file to add path prefix replacement
or deletion.

A comment is now generated in the config file while using the -g or
--generate-config option.

This feature is now used directly in tests to strip the file info
execution dependent paths.

---

Added the following related tests to the `generate_html_outputs`
test-suite:
  - path_prefix_substitution(deletion)
  - path_prefix_substitution(with_substitution_lt_and_gt)
  - path_prefix_substitution(with_substitution_backslash)
  - path_prefix_substitution(with_substitution_backquote)