memcheck_runner.sh: Add support for valgrind's fullpath-after option
Passing the `--fullpath-after` to `memcheck_runner.sh` will forward it
directly to Valgrind.

As stated in valgind's documentation:
By default Valgrind only shows the filenames in stack traces, but not
full paths to source files.
When using Valgrind in large projects where the sources reside in
multiple different directories, this can be inconvenient.
`--fullpath-after` provides a flexible solution to this problem.
When this option is present, the path to each source file is shown,
with the following all-important caveat:
if string is found in the path, then the path up to and including
string is omitted, else the path is shown unmodified.
Note that string is not required to be a prefix of the path.

For example, consider a file named /home/jdoe/blah/src/foo/bar/xyzzy.c.
Specifying --fullpath-after=/home/jdoe/blah/src/ will cause Valgrind to
show the name as foo/bar/xyzzy.c.

Because the string is not required to be a prefix,
--fullpath-after=src/ will produce the same output. This is useful when
the path contains arbitrary machine-generated characters.
For example, the path /my/build/dir/C32A1B47/blah/src/foo/xyzzy can be
pruned to foo/xyzzy using --fullpath-after=/blah/src/.

If you simply want to see the full path, just specify an empty string:
--fullpath-after=. This isn't a special case, merely a logical
consequence of the above rules.

Finally, you can use --fullpath-after multiple times. Any appearance of
it causes Valgrind to switch to producing full paths and applying the
above filtering rule.
Each produced path is compared against all the
--fullpath-after-specified strings, in the order specified. The first
string to match causes the path to be truncated as described above.
If none match, the full path is shown.
This facilitates chopping off prefixes when the sources are drawn from
a number of unrelated directories.